Added controller for testing given the new controller system.
