+++ config/services.yml
@@ -1,0 +1,9 @@
+services:
+
+    example.controller.main:
+        class: phpbb_ext_naderman_example_controller_main
+        arguments:
+            - @request
+            - @user
+            - @template
+            - @dbal.conn
